Pull out rdi instantiation code into separate utility to allow for eventual direct inclusion in Waterline (and mainly just to make it a little easier to work with in isolation for now.
