+++ lib/build-waterline-ontology.js
@@ -12,9 +12,9 @@ var setAssociationColumnTypes = require('./set-association-column-types');
 /**
  * buildWaterlineOntology()
  *
- * Instantiate a Waterline "collection" for each Sails model,
- * then tell Waterline to initialize the ORM and trigger the
- * callback with that fresh new Waterline ontology.
+ * Instantiate a "live" Waterline model instance for each Sails
+ * model definition, then tell Waterline to initialize the ORM
+ * and trigger the callback with that fresh new Waterline ontology.
  *
  * @required {Dictionary} hook
  * @required {SailsApp} sails
