+++ lib/datastore-methods/validate-datastore-connection.js
@@ -8,6 +8,11 @@
 // Validate that the datastore connection is able to perform the requested
 // function based on adapter api version and other checks.
 
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+// TODO: pull the contents of this file into build-registered-datastore-instance
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+
 var _ = require('@sailshq/lodash');
 
 module.exports = function validateDatastoreConnection(options) {
