+++ lib/initialize.js
@@ -12,12 +12,7 @@ var validateDatastoreConfig = require('./validate-datastore-config');
 var validateModelDef = require('./validate-model-def');
 var buildWaterlineOntology = require('./build-waterline-ontology');
 var loadAdapterFromAppDependencies = require('./load-adapter-from-app-dependencies');
-var validateDatastoreConnection = require('./datastore-methods/validate-datastore-connection');
-var leaseConnection = require('./datastore-methods/lease-connection');
-var sendStatement = require('./datastore-methods/send-statement');
-var sendNativeQuery = require('./datastore-methods/send-native-query');
-var runTransaction = require('./datastore-methods/run-transaction');
-
+var buildRegisteredDatastoreInstance = require('./build-registered-datastore-instance');
 
 
 /**
@@ -190,205 +185,13 @@ module.exports = function initialize(hook, sails, done){
           // Validate datastore config.
           var normalizedDatastoreConfig = validateDatastoreConfig(datastoreName, hook, sails);
 
-          // Build our datastore instance
-          // (a dictionary with public methods as well as other important state)
-          var datastoreInstance = {};
-
           // Find the adapter used by this datastore
           var adapter = hook.adapters[normalizedDatastoreConfig.adapter];
 
-          // Get the adapter api version
-          var adapterApiVersion = adapter.adapterApiVersion || 0;
-
-          // Set the normalizedDatastoreConfig properties on the datastore object
-          // for use in the hook.
-          datastoreInstance.internalConfig = normalizedDatastoreConfig;
-
-          // Store the adapter on the internal datastore
-          datastoreInstance.adapter = adapter;
-
-          //  ╔═╗╦ ╦╔═╗╔╦╗╔═╗╔╗╔╔╦╗  ┌┬┐┬ ┬┌─┐  ┌┬┐┌─┐┌┬┐┌─┐┌─┐┌┬┐┌─┐┬─┐┌─┐
-          //  ╠═╣║ ║║ ╦║║║║╣ ║║║ ║    │ ├─┤├┤    ││├─┤ │ ├─┤└─┐ │ │ │├┬┘├┤
-          //  ╩ ╩╚═╝╚═╝╩ ╩╚═╝╝╚╝ ╩    ┴ ┴ ┴└─┘  ─┴┘┴ ┴ ┴ ┴ ┴└─┘ ┴ └─┘┴└─└─┘
-          //  ┬ ┬┬┌┬┐┬ ┬  ┌┬┐┌─┐┌┬┐┌─┐┌─┐┌┬┐┌─┐┬─┐┌─┐  ┌─┐┬ ┬┌┐┌┌─┐┌┬┐┬┌─┐┌┐┌┌─┐
-          //  ││││ │ ├─┤   ││├─┤ │ ├─┤└─┐ │ │ │├┬┘├┤   ├┤ │ │││││   │ ││ ││││└─┐
-          //  └┴┘┴ ┴ ┴ ┴  ─┴┘┴ ┴ ┴ ┴ ┴└─┘ ┴ └─┘┴└─└─┘  └  └─┘┘└┘└─┘ ┴ ┴└─┘┘└┘└─┘
-          // Augment the datastore with functions for working with it directly.
-          // .transaction, .leaseConnection, and .sendStatement methods.
-
-          //  ╦  ╔═╗╔═╗╔═╗╔═╗  ┌─┐┌─┐┌┐┌┌┐┌┌─┐┌─┐┌┬┐┬┌─┐┌┐┌
-          //  ║  ║╣ ╠═╣╚═╗║╣   │  │ │││││││├┤ │   │ ││ ││││
-          //  ╩═╝╚═╝╩ ╩╚═╝╚═╝  └─┘└─┘┘└┘┘└┘└─┘└─┘ ┴ ┴└─┘┘└┘
-          // Add a wrapper for handling the .leaseConnection method
-          datastoreInstance.leaseConnection = function(during, cb) {
-            // Validate the adapter API version and that the datastore is able
-            // to run the lease connection as far as we can tell.
-            try {
-              validateDatastoreConnection({
-                datastoreIdentity: datastoreName,
-                adapterApiVersion: adapterApiVersion,
-                adapter: adapter
-              });
-            } catch (e) {
-              return cb(e);
-            }
-
-            var adapterDatastore = adapter.datastores[datastoreName];
-
-            // Build up options to send to lease connection
-            var options = {
-              datastoreIdentity: datastoreName,
-              driver: adapterDatastore.driver,
-              manager: adapterDatastore.manager,
-              config: adapterDatastore.config,
-              duringFn: during
-            };
-
-            try {
-              return leaseConnection(options, cb);
-            } catch (e) {
-              // Try and return out the cb if the function throws
-              if (_.isFunction(cb)) {
-                return cb(e);
-              }
-              throw e;
-            }
-          };
-
-
-          //  ╔═╗╔═╗╔╗╔╔╦╗  ┌─┐┌┬┐┌─┐┌┬┐┌─┐┌┬┐┌─┐┌┐┌┌┬┐
-          //  ╚═╗║╣ ║║║ ║║  └─┐ │ ├─┤ │ ├┤ │││├┤ │││ │
-          //  ╚═╝╚═╝╝╚╝═╩╝  └─┘ ┴ ┴ ┴ ┴ └─┘┴ ┴└─┘┘└┘ ┴
-          // Add a wrapper for handling the .sendStatement method
-          datastoreInstance.sendStatement = function(statement, dbConnection, cb) {
-            // Validate the adapter API version and that the datastore is able
-            // to run the lease connection as far as we can tell.
-            try {
-              validateDatastoreConnection({
-                datastoreIdentity: datastoreName,
-                adapterApiVersion: adapterApiVersion,
-                adapter: adapter
-              });
-            } catch (e) {
-              return cb(e);
-            }
-
-            var adapterDatastore = adapter.datastores[datastoreName];
-
-            // Build up options to send to send statement
-            var options = {
-              datastoreIdentity: datastoreName,
-              driver: adapterDatastore.driver,
-              manager: adapterDatastore.manager,
-              config: adapterDatastore.config,
-              statement: statement,
-              dbConnection: dbConnection
-            };
-
-            try {
-              return sendStatement(options, cb);
-            } catch (e) {
-              // Try and return out the cb if the sendStatement call throws
-              if (_.isFunction(cb)) {
-                return cb(e);
-              }
-              return e;
-            }
-          };
-
-
-          //  ╔═╗╔═╗╔╗╔╔╦╗  ┌┐┌┌─┐┌┬┐┬┬  ┬┌─┐  ┌─┐ ┬ ┬┌─┐┬─┐┬ ┬
-          //  ╚═╗║╣ ║║║ ║║  │││├─┤ │ │└┐┌┘├┤   │─┼┐│ │├┤ ├┬┘└┬┘
-          //  ╚═╝╚═╝╝╚╝═╩╝  ┘└┘┴ ┴ ┴ ┴ └┘ └─┘  └─┘└└─┘└─┘┴└─ ┴
-          // Add a wrapper for handling the .sendNativeQuery method
-          datastoreInstance.sendNativeQuery = function(nativeQuery, dbConnection, cb) {
-
-
-
-            // TODO: change function signature ^^
-
-            // TODO: move this code in w/ the rest of the implementation:
-            // (because an explicit cb won't always exist)
-            // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-            // Validate the adapter API version and that the datastore is able
-            // to run the lease connection as far as we can tell.
-            try {
-              validateDatastoreConnection({
-                datastoreIdentity: datastoreName,
-                adapterApiVersion: adapterApiVersion,
-                adapter: adapter
-              });
-            } catch (e) {
-              return cb(e);
-            }
-            // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-            var adapterDatastore = adapter.datastores[datastoreName];
-
-            // Build up options to send to send statement
-            var options = {
-              datastoreIdentity: datastoreName,
-              driver: adapterDatastore.driver,
-              manager: adapterDatastore.manager,
-              config: adapterDatastore.config,
-              nativeQuery: nativeQuery,
-              dbConnection: dbConnection
-            };
-
-            try {
-              return sendNativeQuery(options, cb);
-            } catch (e) {
-              // Try and return out the cb if the sendNativeQuery call throws
-              if (_.isFunction(cb)) {
-                return cb(e);
-              }
-              return e;
-            }
-          };
-
-
-          //  ╔╦╗╦═╗╔═╗╔╗╔╔═╗╔═╗╔═╗╔╦╗╦╔═╗╔╗╔
-          //   ║ ╠╦╝╠═╣║║║╚═╗╠═╣║   ║ ║║ ║║║║
-          //   ╩ ╩╚═╩ ╩╝╚╝╚═╝╩ ╩╚═╝ ╩ ╩╚═╝╝╚╝
-          // Add a wrapper for handling the .transaction method
-          datastoreInstance.transaction = function(during, cb) {
-            // Validate the adapter API version and that the datastore is able
-            // to run the lease connection as far as we can tell.
-            try {
-              validateDatastoreConnection({
-                datastoreIdentity: datastoreName,
-                adapterApiVersion: adapterApiVersion,
-                adapter: adapter
-              });
-            } catch (e) {
-              return cb(e);
-            }
-
-            var adapterDatastore = adapter.datastores[datastoreName];
-
-            // Build up options to send to send statement
-            var options = {
-              datastoreIdentity: datastoreName,
-              driver: adapterDatastore.driver,
-              manager: adapterDatastore.manager,
-              config: adapterDatastore.config,
-              duringFn: during
-            };
-
-            try {
-              return runTransaction(options, cb);
-            } catch (e) {
-              // Try and return out the cb if the runTransaction call throws
-              if (_.isFunction(cb)) {
-                return cb(e);
-              }
-              return e;
-            }
-          };//</attach .transaction() function>
-
-
-          // Expose this datastore on `hook.datastores`.
-          hook.datastores[datastoreName] = datastoreInstance;
+          // Build our registered datastore instance (s.k.a. "rdi") - a dictionary
+          // with public methods as well as other important state.  Then expose
+          // our new rdi on `hook.datastores`.
+          hook.datastores[datastoreName] = buildRegisteredDatastoreInstance(datastoreName, normalizedDatastoreConfig, adapter);
 
         });//</_.each() datastore name>
